requirements.txt
flask>=3.0.0
anthropic>=0.84.0
google-generativeai>=0.8.0
Pillow>=10.0.0
python-dotenv>=1.0.0
upstash-redis>=1.0.0
